requirements.txt
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
numpy==1.23.2